+++ src/python/pants/backend/graph_info/tasks/cloc.py
@@ -6,7 +6,7 @@ import os
 from pants.backend.graph_info.subsystems.cloc_binary import ClocBinary
 from pants.base.workunit import WorkUnitLabel
 from pants.engine.fs import FilesContent, PathGlobs, PathGlobsAndRoot
-from pants.engine.isolated_process import Process
+from pants.engine.process import Process
 from pants.task.console_task import ConsoleTask
 from pants.util.contextutil import temporary_dir
