+++ src/python/pants/backend/jvm/subsystems/zinc.py
@@ -20,7 +20,7 @@ from pants.base.build_environment import get_buildroot
 from pants.base.workunit import WorkUnitLabel
 from pants.binaries.binary_tool import NativeTool
 from pants.engine.fs import DirectoryToMaterialize, PathGlobs, PathGlobsAndRoot
-from pants.engine.isolated_process import Process
+from pants.engine.process import Process
 from pants.java.distribution.distribution import Distribution
 from pants.java.jar.jar_dependency import JarDependency
 from pants.util.dirutil import fast_relpath, safe_delete, safe_mkdir, safe_mkdir_for
