+++ src/python/pants/backend/jvm/tasks/jvm_compile/javac/javac_compile.py
@@ -15,7 +15,7 @@ from pants.base.build_environment import get_buildroot
 from pants.base.exceptions import TaskError
 from pants.base.workunit import WorkUnit, WorkUnitLabel
 from pants.engine.fs import DirectoryToMaterialize
-from pants.engine.isolated_process import Process
+from pants.engine.process import Process
 from pants.java.distribution.distribution import DistributionLocator
 from pants.util.dirutil import fast_relpath, safe_walk
 from pants.util.meta import classproperty
