+++ src/python/pants/backend/python/rules/BUILD
@@ -16,7 +16,7 @@ python_library(
     'src/python/pants/engine:addresses',
     'src/python/pants/engine:collection',
     'src/python/pants/engine:fs',
-    'src/python/pants/engine:isolated_process',
+    'src/python/pants/engine:process',
     'src/python/pants/engine:rules',
     'src/python/pants/engine:selectors',
     'src/python/pants/engine:target',
@@ -43,7 +43,7 @@ python_tests(
     'src/python/pants/core/util_rules',
     'src/python/pants/engine:addresses',
     'src/python/pants/engine:fs',
-    'src/python/pants/engine:isolated_process',
+    'src/python/pants/engine:process',
     'src/python/pants/engine:rules',
     'src/python/pants/engine:selectors',
     'src/python/pants/engine/legacy:graph',
