+++ src/python/pants/core/goals/BUILD
@@ -15,7 +15,7 @@ python_library(
     'src/python/pants/engine:fs',
     'src/python/pants/engine:goal',
     'src/python/pants/engine:interactive_runner',
-    'src/python/pants/engine:isolated_process',
+    'src/python/pants/engine:process',
     'src/python/pants/engine:rules',
     'src/python/pants/engine:selectors',
     'src/python/pants/engine:target',
