+++ src/python/pants/core/goals/test.py
@@ -21,7 +21,7 @@ from pants.engine.console import Console
 from pants.engine.fs import Digest, DirectoryToMaterialize, Workspace
 from pants.engine.goal import Goal, GoalSubsystem
 from pants.engine.interactive_runner import InteractiveProcessRequest, InteractiveRunner
-from pants.engine.isolated_process import FallibleProcessResult
+from pants.engine.process import FallibleProcessResult
 from pants.engine.rules import goal_rule, rule
 from pants.engine.selectors import Get, MultiGet
 from pants.engine.target import (
