+++ src/python/pants/goal/context.py
@@ -11,7 +11,7 @@ from pants.base.build_environment import get_buildroot, get_scm
 from pants.base.worker_pool import SubprocPool
 from pants.base.workunit import WorkUnit, WorkUnitLabel
 from pants.build_graph.target import Target
-from pants.engine.isolated_process import (
+from pants.engine.process import (
     FallibleProcessResult,
     ProductDescription,
     fallible_to_exec_result_or_raise,
